+++ packages/sdk/src/contract/contract.factory.ts
@@ -1,5 +1,5 @@
 import { ContractConfig } from '@moonbeam-network/xcm-builder';
-import { Signer } from 'ethers';
+import { EvmSigner } from '../sdk.interfaces';
 import {
   BalanceContractInterface,
   TransferContractInterface,
@@ -8,7 +8,7 @@ import { Erc20, Xtokens } from './contracts';
 
 export function createContract(
   config: ContractConfig,
-  signer: Signer,
+  signer: EvmSigner,
 ): TransferContractInterface | BalanceContractInterface {
   if (config.module === 'Erc20') {
     return new Erc20(config, signer);