+++ packages/sdk/src/getTransferData/getTransferData.utils.ts
@@ -2,14 +2,14 @@ import { CallType, SubstrateQueryConfig } from '@moonbeam-network/xcm-builder';
 import { AssetConfig } from '@moonbeam-network/xcm-config';
 import { Asset } from '@moonbeam-network/xcm-types';
 import { toBigInt } from '@moonbeam-network/xcm-utils';
-import { Signer as EthersSigner } from 'ethers';
 import { BalanceContractInterface, createContract } from '../contract';
 import { PolkadotService } from '../polkadot';
+import { EvmSigner } from '../sdk.interfaces';
 
 export interface GetFeeBalancesParams {
   address: string;
   config: AssetConfig;
-  ethersSigner: EthersSigner;
+  evmSigner: EvmSigner;
   polkadot: PolkadotService;
   asset?: Asset;
 }
@@ -17,7 +17,7 @@ export interface GetFeeBalancesParams {
 export async function getBalance({
   address,
   config,
-  ethersSigner,
+  evmSigner,
   polkadot,
 }: GetFeeBalancesParams) {
   const cfg = config.balance.build({
@@ -29,10 +29,7 @@ export async function getBalance({
     return polkadot.query(cfg as SubstrateQueryConfig);
   }
 
-  const contract = createContract(
-    cfg,
-    ethersSigner,
-  ) as BalanceContractInterface;
+  const contract = createContract(cfg, evmSigner) as BalanceContractInterface;
 
   return contract.getBalance();
 }
@@ -41,7 +38,7 @@ export async function getDecimals({
   address,
   asset,
   config,
-  ethersSigner,
+  evmSigner,
   polkadot,
 }: GetFeeBalancesParams) {
   const cfg = config.balance.build({
@@ -53,10 +50,7 @@ export async function getDecimals({
     return polkadot.getAssetDecimals(asset || config.asset);
   }
 
-  const contract = createContract(
-    cfg,
-    ethersSigner,
-  ) as BalanceContractInterface;
+  const contract = createContract(cfg, evmSigner) as BalanceContractInterface;
 
   return contract.getDecimals();
 }